+++ mocks/mocks.config.js
@@ -2,5 +2,6 @@ module.exports = {
   options: {
     port: 3200,
     log: "debug",
+    pathLegacy: "legacy-mocks",
   },
 };